+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Bug fixes
+
+- Read images on non-Latin Windows paths via `np.fromfile` + `cv2.imdecode` instead of `cv2.imread` (#17).
+
 ## [0.9.1] - 2026-05-26
 
 ### Enhancements